+++ CHANGELOG.md
@@ -94,3 +94,4 @@
 * Rename `idle` method of `Avoidable` typeclass to `empty`
 * Remove `a` parameter from `Layout` to be able to use natural transformations in methods
 * Return `filter` method to `Stack` data structure
+* Rename `unwrap` to `untwist` method in `Twister` module